LayoutTests/fast/js/kde/string-2-n-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.successfullyParsed is undefined.

TEST COMPLETE
